Move calculate functions to calculate.py
