+++ agent/basic_pokemon_agent.py
@@ -1,13 +1,13 @@
 """Class for a pokemon player."""
 
 from numpy.random import uniform
+
 from agent.base_agent import BaseAgent
+from config import POKEMON_DATA
+from pokemon_helpers.calculate import calculate_spe_range
+from pokemon_helpers.calculate import generate_all_ev_combinations
 from pokemon_helpers.damage_stats import DamageStatCalc
 from pokemon_helpers.pokemon import Pokemon
-from pokemon_helpers.pokemon import calculate_spe_range
-from pokemon_helpers.pokemon import generate_all_ev_combinations
-
-from config import POKEMON_DATA
 
 
 class PokemonAgent(BaseAgent):
